Separeted logics and styles from Nav component
