+++ src/components/nav/useNav.ts
@@ -1,0 +1,37 @@
+import { useState } from "react";
+
+export const useNav = () => {
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
+  const [mobileMoreAnchorEl, setMobileMoreAnchorEl] =
+    useState<null | HTMLElement>(null);
+
+  const isMenuOpen = Boolean(anchorEl);
+  const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
+
+  const handleProfileMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleMobileMenuClose = () => {
+    setMobileMoreAnchorEl(null);
+  };
+
+  const handleMenuClose = () => {
+    setAnchorEl(null);
+    handleMobileMenuClose();
+  };
+
+  const handleMobileMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
+    setMobileMoreAnchorEl(event.currentTarget);
+  };
+  return {
+    anchorEl,
+    mobileMoreAnchorEl,
+    isMenuOpen,
+    isMobileMenuOpen,
+    handleProfileMenuOpen,
+    handleMobileMenuClose,
+    handleMobileMenuOpen,
+    handleMenuClose,
+  };
+};
